feat: Implement comprehensive installer, logging, and startup fixes
- **Installer:**
  - Implemented a WiX Burn bootstrapper to create an `.exe` installer.
  - The installer now checks for the .NET 9 Desktop Runtime and prompts the user to download and install it if it's missing.
  - Added a feature tree UI to the installer to allow users to choose whether to create desktop and Start Menu shortcuts.
  - The installer now creates a custom directory structure, including a `lib` folder for DLLs.

- **Bug Fixes:**
  - Fixed a critical startup crash by adding an `App.config` file with a probing path to the `lib` directory.
  - Corrected the `ConfigurationService` and `LightJockeyEntertainmentConfig` to properly handle secure and non-secure settings without breaking the build.
  - Fixed numerous build errors and warnings.
  - Refactored `ViewModelBase` to correctly use `ObservableObject` from the MVVM Toolkit.

- **Features:**
  - Implemented a configurable logging system with a new settings window.
  - Users can now configure the log level and the number of log files to retain.
  - Log files are now created with a timestamp in their name.

+++ src/LightJockey/Services/EntertainmentService.cs
@@ -319,7 +319,7 @@ public class EntertainmentService : IEntertainmentService
         try
         {
             // Start auto-update for the streaming group
-            _ = _streamingClient.AutoUpdateAsync(_streamingGroup, cancellationToken, updateRateMs, onlySendDirtyStates: false);
+            _ = _streamingClient.AutoUpdateAsync(_streamingGroup, cancellationToken, (int)updateRateMs, onlySendDirtyStates: false);
 
             while (!cancellationToken.IsCancellationRequested)
             {
